chore: rename able/applications directory
