+++ packages/frontend/src/pages/Applications.svelte
@@ -4,13 +4,13 @@ import NavPage from '../lib/NavPage.svelte';
 import Table from '../lib/table/Table.svelte';
 import { Column, Row } from '../lib/table/table';
 import { applicationStates } from '../stores/application-states';
-import ColumnActions from '../lib/table/environment/ColumnActions.svelte';
-import ColumnStatus from '../lib/table/environment/ColumnStatus.svelte';
+import ColumnActions from '../lib/table/application/ColumnActions.svelte';
+import ColumnStatus from '../lib/table/application/ColumnStatus.svelte';
 import type { ApplicationCell } from './applications';
-import ColumnRecipe from '../lib/table/environment/ColumnRecipe.svelte';
-import ColumnModel from '../lib/table/environment/ColumnModel.svelte';
-import ColumnPod from '../lib/table/environment/ColumnPod.svelte';
-import ColumnAge from '../lib/table/environment/ColumnAge.svelte';
+import ColumnRecipe from '../lib/table/application/ColumnRecipe.svelte';
+import ColumnModel from '../lib/table/application/ColumnModel.svelte';
+import ColumnPod from '../lib/table/application/ColumnPod.svelte';
+import ColumnAge from '../lib/table/application/ColumnAge.svelte';
 import { filterByLabel } from '/@/utils/taskUtils';
 import { tasks } from '/@/stores/tasks';
